local-app/requirements.txt
nfcpy==1.0.4
requests==2.31.0
python-dotenv==1.0.0
schedule==1.1.0
